src/ml/requirements.txt
boto3
scikit-learn
pandas
numpy
python-dateutil
